Netplay: Ok After more testing it doesnt seem to be the CDROM. so im reveting this change. its weird tho its seems to be fine in the first session created but after when a new session is created it is prone to desync. this doesnt happen when starting the session with an identical savestate for both players tho. GGPO: Changed the size of the checksums from u16 to u32.

+++ dep/ggpo-x/include/ggponet.h
@@ -196,8 +196,8 @@ typedef struct {
       } chat;
       struct {
           int nFrameOfDesync;
-          uint16_t ourCheckSum;
-          uint16_t remoteChecksum;
+          uint32_t ourCheckSum;
+          uint32_t remoteChecksum;
       } desync;
    } u;
 } GGPOEvent;
